+++ src/constants/index.ts
@@ -7,3 +7,4 @@ export * from './storage.constants';
 export * from './webWalletProvider.constants';
 export * from './window.constants';
 export * from './websocket.constants';
+export * from './providerFactory.constants';
