+++ src/core/providers/types/providerFactory.types.ts
@@ -34,16 +34,6 @@ export enum ProviderTypeEnum {
   none = ''
 }
 
-export const providerLabels: Record<string, string> = {
-  [ProviderTypeEnum.crossWindow]: 'MultiversX Web Wallet',
-  [ProviderTypeEnum.extension]: 'MultiversX Wallet Extension',
-  [ProviderTypeEnum.walletConnect]: 'xPortal App',
-  [ProviderTypeEnum.ledger]: 'Ledger',
-  [ProviderTypeEnum.metamask]: 'MetaMask Snap',
-  [ProviderTypeEnum.passkey]: 'Passkey',
-  [ProviderTypeEnum.none]: ''
-};
-
 export interface IProviderFactory<
   T extends ProviderTypeEnum = ProviderTypeEnum
 > {
